restore chainconfig as separate module, move config under service
