+++ fuel-core/src/chain_config/serialization.rs
