+++ fuel-core/src/lib.rs
@@ -1,5 +1,5 @@
+pub mod chain_config;
 pub mod coin_query;
-pub mod config;
 pub mod database;
 pub mod executor;
 pub mod model;
