+++ fuel-core/src/schema/chain.rs
@@ -1,6 +1,6 @@
 use crate::{
-    config::Config, database::Database, model::FuelBlockDb, schema::block::Block,
-    schema::scalars::U64,
+    database::Database, model::FuelBlockDb, schema::block::Block, schema::scalars::U64,
+    service::Config,
 };
 use async_graphql::{Context, Object};
 use fuel_core_interfaces::common::{fuel_storage::Storage, fuel_tx, fuel_types};
