+++ fuel-core/src/schema/tx.rs
@@ -1,8 +1,8 @@
-use crate::config::Config;
 use crate::database::{transaction::OwnedTransactionIndexCursor, Database, KvStoreError};
 use crate::executor::Executor;
 use crate::model::{BlockHeight, FuelBlockDb};
 use crate::schema::scalars::{Address, Bytes32, HexString, SortedTxCursor, TransactionId};
+use crate::service::Config;
 use crate::state::IterDirection;
 use anyhow::anyhow;
 use async_graphql::{
