+++ fuel-core/src/service.rs
@@ -1,7 +1,4 @@
-use crate::{
-    config::{Config, DbType},
-    database::Database,
-};
+use crate::database::Database;
 use anyhow::Error as AnyError;
 use modules::Modules;
 use std::{net::SocketAddr, panic};
@@ -9,6 +6,9 @@ use thiserror::Error;
 use tokio::task::JoinHandle;
 use tracing::log::warn;
 
+pub use config::{Config, DbType};
+
+pub mod config;
 pub(crate) mod genesis;
 pub mod graph_api;
 pub mod metrics;
