+++ fuel-core/src/service/config.rs
@@ -1,12 +1,10 @@
-pub mod chain_config;
-pub mod serialization;
-
-use chain_config::ChainConfig;
+use crate::chain_config::ChainConfig;
 use std::{
     net::{Ipv4Addr, SocketAddr},
     path::PathBuf,
 };
 use strum_macros::{Display, EnumString, EnumVariantNames};
+
 #[derive(Clone, Debug)]
 pub struct Config {
     pub addr: SocketAddr,
