+++ fuel-tests/tests/node_info.rs
@@ -1,4 +1,4 @@
-use fuel_core::{config::Config, service::FuelService};
+use fuel_core::service::{Config, FuelService};
 use fuel_gql_client::client::{schema::node_info::NodeInfo, FuelClient};
 
 #[tokio::test]
